Chat_bot/requirements.txt
Flask==2.3.3
llama-cpp-python==0.2.20
jquery
